Specify check parameters with CLI arguments
`--version` sets the Tarantool version to test, such as 1.10 or 2.10
`--build` sets the Tarantool build, as it appears in the Downloads
  API: script, manual or nightly.
`--dist` -- the OS distribution to test on, such as 'debian'
`--dist-version` -- the exact version of the OS distribution above,
  such as '11' or 'bullseye'

With either of these arguments present, the actual list of checked
distributions is filtered from what's returned by the website's API
and what's found in the JSON config.

For example, API can return the following list for CentOS builds:

- centos_script_2.10
- centos_manual_2.10
- centos_script_1.10
- centos_manual_1.10

With `--version 2.10 --build script`, the Checker will only check
the script build instructions for Tarantool 2.10 on all versions
of CentOS, available in the JSON config, as well as on other
distributions and their version.

Adding `--dist centos --dist-version 7` will narrow the check to CentOS 7 only.

That is, to check the script (default) installation of Tarantool 2.10 on
CentOS 7:

```python
python check.py --version 2.10 --build script --dist centos --dist-version 7
```

Part of #45