Check installation of GC64 packages
Check the installation of GC64 packages for the x86_64 platform, for
which we have separate repositories with such packages.

There is no need to test the installation of GC64 packages for the
aarch64 platform. All packages for this platform are GC64 ones by
default. That's why we don't even have separate repositories for them.

Note, for now we cannot test the manual instruction for the installation
of GC64 packages because it is not implemented on the site.

Closes #39